chore: declare types and move old imports to utils and constants
