+++ lib/src/components/Select/docs/examples/clearable.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
-import type { SelectProps } from '@/components/Select'
 import { Select } from '@/components/Select'
+import type { SelectProps } from '@/components/Select/types'
 
 export const ITEMS = [
   { label: 'Bold', value: 'bold' },
