+++ lib/src/components/Select/docs/examples/formatting.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 
 import { WttjIcon } from '@/components/Icon'
-import type { SelectOption, SelectProps } from '@/components/Select'
 import { Select } from '@/components/Select'
+import type { SelectOption, SelectProps } from '@/components/Select/types'
 
 export const ITEMS = [
   { label: 'Bold', value: 'bold' },
