+++ lib/src/components/Select/docs/examples/icon.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 
 import { WttjIcon } from '@/components/Icon'
-import type { SelectProps } from '@/components/Select'
 import { Select } from '@/components/Select'
+import type { SelectProps } from '@/components/Select/types'
 
 export const ITEMS = [
   { label: 'Bold', value: 'bold' },
