+++ lib/src/components/Select/docs/examples/option-groups.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
-import type { SelectProps } from '@/components/Select'
 import { Select } from '@/components/Select'
+import type { SelectProps } from '@/components/Select/types'
 import { Box } from '@old/Box'
 import { Tag } from '@old/Tag'
 import { Text } from '@old/Text'
