+++ lib/src/types.ts
@@ -1,0 +1,7 @@
+// we want to keep DefaultFieldIconSize in a natural order for documentation
+// eslint-disable-next-line perfectionist/sort-union-types
+export type DefaultFieldIconSize = 'xs' | 'sm'
+
+// we want to keep Size in a natural order for documentation
+// eslint-disable-next-line perfectionist/sort-union-types
+export type Size = 'xs' | 'sm' | 'md' | 'lg'
